Refactor: Move age group utility functions to populations.py
